fixed: when switching games ini settings could be lost
when switching away from a profile with local settings, the current intended behavior is for vortex to copy the ini files from the location where the game uses them to the profile, this step failed.
Unfortunately the return path where upon activating the profile the profile in was copied to the global location did work and thus reset the ini.